added support to ignore specific journal entries.
This will make it easier for those WTF entries that happen due to various transfers or administrative overhead. These entries tend to screw with stuff,
and their value of worth cannot be calculated by a script. It's up to the corp leadership to maintain the ignore list, selecting what is not important
and what is.

Ignored entries are stored in a plaintext .cfg file in the same directory.